% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2d3.R
\name{d3_rollcall_idealpoints}
\alias{d3_rollcall_idealpoints}
\title{D3.js Ideal Point Estimation for legislative roll call data}
\usage{
d3_rollcall_idealpoint()
}
\arguments{
\item{data}{= data}

\item{width}{= NULL}

\item{height}{= NULL}
}
\description{
This function creates a D3.js ideal point estimation graph for legislative roll call data.
}
\section{Building upon R ideal points graphs}{

The function wraps around D3.js and requires installation of the RStudio daily build (see https://rstudio.github.io/r2d3/ for clarification). You can set the width and height of the graph. You must call the data by calling a data frame or matrix.
}

\examples{
d3_rollcall_idealpoints(data=estimates, height=10, width=6)
}
\author{
Zoe Meers, Simon Jackman
}
