% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ussc_logo.R
\name{plot_ussc_logo}
\alias{plot_ussc_logo}
\title{Append new USSC logo at the bottom of graph}
\usage{
plot_ussc_logo()
}
\arguments{
\item{position}{A character vector of the logo position - "bottom-right", "bottom-centre","bottom-center","bottom-left","top-right", "top-centre","top-center","top-left".}
}
\description{
Append new USSC logo at the bottom of graph
}
\author{
Zoe Meers
}
