% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{prop_grouped_survey_question}
\alias{prop_grouped_survey_question}
\title{Grouped proportional tables for USSC cross-country surveys}
\usage{
prop_grouped_survey_question()
}
\description{
Calculates the weighted proportion for the specified questions and groups, both in the US and Australia.
}
\examples{
survey_data \%>\% 
prop_grouped_survey_question(questions = starts_with("sexual_harrassment"), partyid, gender)
}
\author{
Zoe Meers
}
