% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{prop_survey_question}
\alias{prop_survey_question}
\title{Prop tables for USSC cross-country surveys}
\usage{
prop_survey_question()
}
\description{
Calculates the proportion for the specified questions in the US and Australia.
}
\examples{
survey_data \%>\% 
prop_survey_question(questions = starts_with("importance_of"))
}
\author{
Zoe Meers
}
