% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_tables.R
\name{relevel_survey_answer}
\alias{relevel_survey_answer}
\title{Relevel answer column as factors}
\usage{
relevel_survey_answer()
}
\description{
Relevel answer column as factors
}
\examples{
\dontrun{
survey_data \%>\% 
prop_survey_question(questions = starts_with("importance_of")) \%>\% 
relevel_survey_answer(levels)
}
}
\author{
Zoe Meers
}
