% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVG_Halis_GR.R
\name{save_svg_fonts}
\alias{save_svg_fonts}
\title{SVG files with Halis GR fonts}
\usage{
save_svg_fonts()
}
\arguments{
\item{load_file}{Load old file that needs to be edited to host the USSC fonts.}

\item{save_file}{Save new file to directory.}
}
\description{
Ensures USSC fonts will always load in SVG format, regardless of whether user has the Halis GR fonts installed on their local machine.
To do this, the person creating the viz must save the SVG object and then reload it, saving it again using this function.
}
\examples{
\dontrun{
save_svg_fonts(load_file = "file.svg", save_file = "new_file.svg")
}
}
\author{
Zoe Meers
}
