% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\docType{data}
\name{theme_ussc}
\alias{theme_ussc}
\title{Main USSC theme}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 58.}
\usage{
theme_ussc
}
\description{
Main USSC theme
}
\examples{
Create ggplot theme using main USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc + labs(title="Neo Sans Pro Header", x="Univsers Font: Sepal Width", y="Univers Font: Sepal Length") + scale_colour_ussc("blue")
}
\keyword{datasets}
