% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{theme_ussc_dark}
\alias{theme_ussc_dark}
\title{Dark USSC theme}
\usage{
theme_ussc_dark()
}
\description{
Calls the dark USSC theme.
}
\examples{
# Create ggplot graph using dark USSC theme
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Width, Sepal.Length, colour = Species)) + 
ggplot2::geom_point(size = 4, alpha=0.4) +  
ussc::theme_ussc_dark() + 
ggplot2::labs(title="Neo Sans Pro Header", x="Univers Font: Sepal Width", y="Univers Font: Sepal Length") + 
ussc::scale_colour_ussc("blue")
}
\author{
Zoe Meers
}
