% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\docType{data}
\name{theme_ussc_neosanspro}
\alias{theme_ussc_neosanspro}
\title{USSC theme with Neo Sans Pro font}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 58.}
\usage{
theme_ussc_neosanspro
}
\description{
USSC theme with Neo Sans Pro font
}
\examples{
Create ggplot theme using Neo Sans Pro USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc_neosanspro + labs(title="Neo Sans Pro Header", x="Sepal Width", y="Sepal Length") + scale_color_ussc()
}
\keyword{datasets}
