% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\docType{data}
\name{theme_ussc_univers}
\alias{theme_ussc_univers}
\title{USSC theme with Univers font}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 58.}
\usage{
theme_ussc_univers
}
\description{
USSC theme with Univers font
}
\examples{
Create ggplot theme using Univers USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc_univers + labs(title="Univers Header", x="Sepal Width", y="Sepal Length") + scale_colour_ussc()
}
\keyword{datasets}
