% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\name{ussc_colors}
\alias{ussc_colors}
\title{Function to extract USSC colours as hex codes}
\usage{
ussc_colors(...)
}
\arguments{
\item{...}{Character names of ussc_colors}
}
\description{
Function to extract USSC colours as hex codes
}
\author{
Zoe Meers
}
