% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\name{ussc_colours}
\alias{ussc_colours}
\title{Function to extract USSC colours as hex codes}
\usage{
ussc_colours(...)
}
\arguments{
\item{...}{Character names of ussc_colours}
}
\description{
Function to extract USSC colours as hex codes
}
\examples{
ussc::ussc_colours(c("light blue", "orange"))
}
\author{
Zoe Meers
}
