% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_confluence.R
\name{ussc_confluence_word_tables}
\alias{ussc_confluence_word_tables}
\title{USSC Confluence Word Files & Tables}
\usage{
ussc_confluence_word_tables(id = id,
  username = Sys.getenv("CONFLUENCE_USERNAME"),
  password = Sys.getenv("CONFLUENCE_PASSWORD"))
}
\arguments{
\item{id}{Page ID - a number found in the confluence URL}

\item{username}{Your Confluence username which should be identical to your email. Defaults to an entry in .renviron file called CONFLUENCE_USERNAME.}

\item{password}{Your Confluence API key (Get from https://confluence.atlassian.com/cloud/api-tokens-938839638.html). Defaults to an entry in .renviron file called CONFLUENCE_PASSWORD.}
}
\description{
USSC Confluence Word Files & Tables
}
\examples{
ussc_confluence_word_tables(id = "30441548")
}
\author{
Zoe Meers
}
