% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{ussc_fonts}
\alias{ussc_fonts}
\title{Loads USSC fonts into R from the local font book. Download necessary fonts from the fonts folder in this package and add to font book (if you have not done so before).}
\usage{
ussc_fonts()
}
\description{
Loads USSC fonts into R from the local font book. Download necessary fonts from the fonts folder in this package and add to font book (if you have not done so before).
}
