% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_google_search.R
\name{ussc_google_create_custom_query}
\alias{ussc_google_create_custom_query}
\title{Creates custom Google Search query}
\usage{
ussc_google_create_custom_query()
}
\arguments{
\item{.data}{The data frame containing the strings you want to turn into a query.}

\item{...}{A list of columns to convert -- this can just be one column.}
}
\description{
Defines the query needed for Google Search to grab the right data.
}
\examples{
df <- tibble::tribble(~name, ~place, "Simon Jackman", "USSC")
ussc::ussc_google_create_custom_query(data = df, "name", "place") 
}
\author{
Zoe Meers
}
