% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_google_search.R
\name{ussc_google_total_results}
\alias{ussc_google_total_results}
\title{Grabs data from Google Search API}
\usage{
ussc_google_total_results()
}
\arguments{
\item{.data}{data frame with custom queries}

\item{api_key}{Google Search API -- follow the instructions here https://developers.google.com/custom-search/v1/overview#api_key.}

\item{cx_id}{Custom Search Engine ID
Once you have an api_key, create a Custom Search Engine. 
Follow the instructions here https://developers.google.com/custom-search/docs/tutorial/creatingcse#defining_a_custom_search_engine_in_control_panel. 
In the form, for the website field type in www.google.com. 
Remember to give the Custom Search Engine an informative name. 
Now you need to edit the CSE to search for all links (not just www.google.com). 
Go here: https://cse.google.com/create/new. On the left, click edit search engine, 
click on the CSE that you just created and scroll down to "Search the entire web". Click Yes. 
Scroll up, copy Search engine ID, paste in your .renviron file under GOOGLE_SEARCH_CX_ID.}

\item{query}{Google Search query from ussc::ussc_google_create_custom_query}
}
\description{
Uses the google search query to grab total search numbers from Google Search.
}
\examples{
\dontrun{
df <- tibble::tribble(~name, ~place, "Simon Jackman", "USSC")
custom_query <- ussc::ussc_google_create_custom_query(df, "name", "place") 
results <- ussc::ussc_google_total_results(query = custom_query)
}
}
\author{
Zoe Meers
}
