% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\name{ussc_palettes}
\alias{ussc_palettes}
\title{Return function to interpolate a ussc colour palette}
\usage{
ussc_palettes(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ussc_palettes (i.e. main, blue, light, dark, grey, mixed)}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a ussc colour palette
}
\author{
Zoe Meers
}
