% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\docType{data}
\name{ussc_palettes_list}
\alias{ussc_palettes_list}
\title{USSC palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
ussc_palettes_list
}
\description{
This function prints the `ussc` colors as organised by political party or dark or light color scheme.
}
\examples{
ussc::ussc_palettes_list
}
\author{
Zoe Meers
}
\keyword{datasets}
