% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ussc_logo.R
\name{plot_ussc_logo}
\alias{plot_ussc_logo}
\title{Append new USSC logo at the bottom of graph}
\usage{
plot_ussc_logo()
}
\description{
Append new USSC logo at the bottom of graph
}
\examples{
plot_ussc_logo(ggplot_object)
}
\author{
Zoe Meers
}
