% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\name{scale_fill_ussc}
\alias{scale_fill_ussc}
\title{Fill scale constructor for USSC colours}
\usage{
scale_fill_ussc(palette = "main", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in ussc_pal (i.e. main, blue, light, dark, grey, mixed)}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradient(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill scale constructor for USSC colours
}
\examples{
Fill by discrete variable with different palette + remove legend (guide)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) + geom_bar() + theme(axis.text.x = element_text(angle = 45, hjust = 1)) + scale_fill_ussc(palette = "main", guide = "none")
Set colour transparency by calling alpha in the geom_...() command
ggplot(mpg, aes(manufacturer, fill = manufacturer)) + geom_bar(alpha=0.7) + theme(axis.text.x = element_text(angle = 45, hjust = 1)) + scale_fill_ussc(palette = "main", guide = "none")
}
\author{
Zoe Meers
}
