% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{theme_ussc}
\alias{theme_ussc}
\title{Main USSC theme}
\usage{
theme_ussc()
}
\description{
Calls the main USSC theme. A reconstructed theme_bw() ggplot2 theme.
}
\examples{
Create ggplot theme using main USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc() + labs(title="Univers Header", x="Univers Font: Sepal Width", y="Univers Font: Sepal Length") + scale_colour_ussc("blue")
}
\author{
Zoe Meers
}
