% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{theme_ussc}
\alias{theme_ussc}
\title{Main USSC theme}
\usage{
theme_ussc()
}
\description{
Calls the main USSC theme.
}
\examples{
# Create ggplot graph using main USSC theme
\dontshow{suppressWarnings(}
ggplot2::ggplot(iris, ggplot2::aes(Sepal.Width, Sepal.Length, colour = Species)) + 
ggplot2::geom_point(size = 4, alpha=0.4) + 
ussc::theme_ussc() + 
ggplot2::labs(title="Halis GR Header", x="Sepal Width", y="Sepal Length") + 
ussc::scale_colour_ussc("blue")
\dontshow{)}
}
\author{
Zoe Meers
}
