% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{theme_ussc_neosanspro_light}
\alias{theme_ussc_neosanspro_light}
\title{USSC theme with light Neo Sans Pro font}
\usage{
theme_ussc_neosanspro_light()
}
\description{
Theme_minimal() with light Neo Sans Pro font
}
\examples{
Create ggplot theme using light Neo Sans Pro USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc_neosanspro_light + labs(title="Light Neo Sans Pro Header", x="Sepal Width", y="Sepal Length") + scale_colour_ussc()
}
\author{
Zoe Meers
}
