% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ussc.R
\name{theme_ussc_univers}
\alias{theme_ussc_univers}
\title{USSC theme with Univers font}
\usage{
theme_ussc_univers()
}
\description{
A theme_minimal() with Univers font
}
\examples{
Create ggplot theme using Univers USSC theme
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) + geom_point(size = 4, alpha=0.4) +  theme_ussc_univers() + labs(title="Univers Header", x="Sepal Width", y="Sepal Length") + scale_colour_ussc()
}
\author{
Zoe Meers
}
