% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\docType{data}
\name{ussc_pal}
\alias{ussc_pal}
\title{USSC palettes}
\format{An object of class \code{list} of length 4.}
\usage{
ussc_pal
}
\description{
This list creates palette types.
}
\examples{
ggplot2::ggplot(mtcars, aes(hp, mpg)) + ggplot2::geom_point(colour = ussc_colours("orange"), size = 4)
}
\author{
Zoe Meers
}
\keyword{datasets}
