% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ussc_colours.R
\docType{data}
\name{ussc_pal}
\alias{ussc_pal}
\title{USSC palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
ussc_pal
}
\description{
This function prints the `ussc` colors as organised by political party or dark or light color scheme.
}
\examples{
ussc::ussc_pal
}
\author{
Zoe Meers
}
\keyword{datasets}
